refactor: move spinner to core
